Extract ShellCompletionEngine from CoreReplApp.ShellCompletion partial
Move shell completion candidate resolution into a standalone
ShellCompletionEngine class. Promote shared utility methods
(IsGlobalOptionToken, TokenizeInputSpans, CollectVisibleMatchingRoutes)
to internal visibility for cross-class access.
